Try harder to preser worker.log.

+++ janitor/gcb_worker.py
@@ -107,19 +107,21 @@ async def run_gcb_worker(logf, output_directory, args, timeout=None):
         ops = json.loads(r.data.decode('utf-8'))
         if ops['status'] != 'WORKING':
             break
-    assert ops['status'] == 'SUCCESS', ops['status']
-    artifact_manifest_url = ops['results']['artifactManifest']
-    download_results(
-        http, bearer, artifact_manifest_url, output_directory)
     blob = get_blob(http, bearer, ops['logsBucket'] + '/log-%s.txt' % build_id)
     logf.write(blob.data)
-    tgz_name = os.environ['PACKAGE'] + '.tgz'
-    tgz_path = os.path.join(output_directory, tgz_name)
-    if os.path.exists(tgz_path):
-        subprocess.check_call(
-            ['tar', 'xfz', tgz_name],
-            cwd=output_directory)
-        os.unlink(tgz_path)
+    if ops['status'] == 'SUCCESS':
+        artifact_manifest_url = ops['results']['artifactManifest']
+        download_results(
+            http, bearer, artifact_manifest_url, output_directory)
+        tgz_name = os.environ['PACKAGE'] + '.tgz'
+        tgz_path = os.path.join(output_directory, tgz_name)
+        if os.path.exists(tgz_path):
+            subprocess.check_call(
+                ['tar', 'xfz', tgz_name],
+                cwd=output_directory)
+            os.unlink(tgz_path)
+    else:
+        raise AssertionError('build failed with %r' % ops['status'])
 
 
 def main(argv=None):
